+++ bin/release.dart
@@ -1,7 +1,8 @@
 import "dart:convert";
 import "dart:io";
 
-import "helper/copy.dart";
+import "package:path/path.dart" as path;
+import "package:pubspec_parse/pubspec_parse.dart";
 
 Future<void> main(List<String> args) async {
   if (args.isEmpty) {
@@ -16,30 +17,29 @@ Future<void> main(List<String> args) async {
     exit(1);
   }
 
-  // Get current build name and number from pubspec.yaml
-  final pubspec = File("pubspec.yaml");
-  final pubspecContent = await pubspec.readAsString();
+  final pubspec = File("pubspec.yaml").readAsStringSync();
+  final parsed = Pubspec.parse(pubspec);
 
-  // 1.0.0+1
-  final buildRegExp =
-      RegExp(r"version: (.+)").firstMatch(pubspecContent)!.group(1);
+  /// Only base version 1.0.0
+  final buildName =
+      "${parsed.version?.major}.${parsed.version?.minor}.${parsed.version?.patch}";
+  final buildNumber = parsed.version?.build.firstOrNull.toString();
 
-  if (buildRegExp == null) {
-    print("version not found in pubspec.yaml");
-    exit(1);
-  }
+  print(
+    "Building version $buildName+$buildNumber for $platform for app ${parsed.name}",
+  );
 
-  print("Building version ${buildRegExp.replaceAll('"', '')}");
-
-  final buildName = buildRegExp.replaceAll('"', "").split("+").first.trim();
-  final buildNumber = buildRegExp.replaceAll('"', "").split("+").last.trim();
-  final appNamePubspec =
-      RegExp(r"name: (.+)").firstMatch(pubspecContent)!.group(1);
+  final appNamePubspec = parsed.name;
 
   // Get flutter path
   final flutterPath = Platform.environment["FLUTTER_ROOT"];
 
-  // print current working directory
+  if (flutterPath == null || flutterPath.isEmpty) {
+    print("FLUTTER_ROOT environment variable is not set");
+    exit(1);
+  }
+
+  // Print current working directory
   print("Current working directory: ${Directory.current.path}");
 
   // Determine the Flutter executable based on the platform
@@ -47,12 +47,16 @@ Future<void> main(List<String> args) async {
   if (Platform.isWindows) {
     flutterExecutable += ".bat";
   }
-  if (Platform.isLinux || Platform.isMacOS) {
-    flutterExecutable = "flutter";
+
+  final flutterBinPath = path.join(flutterPath, "bin", flutterExecutable);
+
+  if (!File(flutterBinPath).existsSync()) {
+    print("Flutter executable not found at path: $flutterBinPath");
+    exit(1);
   }
 
   final buildCommand = [
-    "$flutterPath${Platform.pathSeparator}bin${Platform.pathSeparator}$flutterExecutable",
+    flutterBinPath,
     "build",
     platform,
     "--dart-define",
@@ -61,12 +65,13 @@ Future<void> main(List<String> args) async {
     "FLUTTER_BUILD_NUMBER=$buildNumber",
   ];
 
+  print("Executing build command: ${buildCommand.join(' ')}");
+
   // Replace Process.run with Process.start to handle real-time output
   final process =
       await Process.start(buildCommand.first, buildCommand.sublist(1));
 
   process.stdout.transform(utf8.decoder).listen(print);
-
   process.stderr.transform(utf8.decoder).listen((data) {
     stderr.writeln(data);
   });
@@ -77,44 +82,56 @@ Future<void> main(List<String> args) async {
     exit(1);
   }
 
-  print("Build completed");
+  print("Build completed successfully");
 
   late Directory buildDir;
 
-  // Found executable file name in build folder
+  // Determine the build directory based on the platform
   if (platform == "windows") {
     buildDir = Directory(
-      "build${Platform.pathSeparator}windows${Platform.pathSeparator}x64${Platform.pathSeparator}runner${Platform.pathSeparator}Release",
+      path.join("build", "windows", "x64", "runner", "Release"),
     );
   } else if (platform == "macos") {
     buildDir = Directory(
-      "build/macos/Build/Products/Release/$appNamePubspec.app",
+      path.join(
+        "build",
+        "macos",
+        "Build",
+        "Products",
+        "Release",
+        "$appNamePubspec.app",
+      ),
     );
   } else if (platform == "linux") {
     buildDir = Directory(
-      "build/linux/x64/release/bundle/",
+      path.join("build", "linux", "x64", "release", "bundle"),
     );
   }
 
-  // final files = await buildDir.list(recursive: true).toList();
-
-  // Get only last part of the path
-  final appName = appNamePubspec;
-
-  late String distPath;
-
-  if (platform == "windows") {
-    distPath =
-        "dist${Platform.pathSeparator}$buildNumber${Platform.pathSeparator}$appName-$buildName+$buildNumber-$platform";
-  } else if (platform == "macos") {
-    distPath =
-        "dist/$buildNumber/$appName-$buildName+$buildNumber-$platform/$appName.app";
-  } else if (platform == "linux") {
-    distPath =
-        "dist/$buildNumber/$appName-$buildName+$buildNumber-$platform";
+  if (!buildDir.existsSync()) {
+    print("Build directory not found: ${buildDir.path}");
+    exit(1);
   }
 
-  // Copy buildDir to distPath, included directory name
+  final distPath = platform == "windows"
+      ? path.join(
+          "dist",
+          buildNumber,
+          "$appNamePubspec-$buildName+$buildNumber-$platform",
+        )
+      : platform == "macos"
+          ? path.join(
+              "dist",
+              buildNumber,
+              "$appNamePubspec-$buildName+$buildNumber-$platform",
+              "$appNamePubspec.app",
+            )
+          : path.join(
+              "dist",
+              buildNumber,
+              "$appNamePubspec-$buildName+$buildNumber-$platform",
+            );
+
   final distDir = Directory(distPath);
   if (distDir.existsSync()) {
     distDir.deleteSync(recursive: true);
@@ -123,5 +140,21 @@ Future<void> main(List<String> args) async {
   // Copy buildDir to distPath
   await copyDirectory(buildDir, Directory(distPath));
 
-  print("Archive created to $distPath");
+  print("Archive created at $distPath");
+}
+
+// Helper function to copy directories recursively
+Future<void> copyDirectory(Directory source, Directory destination) async {
+  if (!destination.existsSync()) {
+    destination.createSync(recursive: true);
+  }
+
+  await for (final entity in source.list(recursive: true)) {
+    if (entity is File) {
+      final relativePath = path.relative(entity.path, from: source.path);
+      final newPath = path.join(destination.path, relativePath);
+      await Directory(path.dirname(newPath)).create(recursive: true);
+      await entity.copy(newPath);
+    }
+  }
 }